Send story chapters to the front end
i get the root node of every story and do a BFS iteration through it, and store it in a dictionary by depth. These lists in the dictionary are ordered in the same way as the angle lists, so you just need to access the same index as you do for the angle.  passing the dictionary to the front end as json.